refactored indivdual statement
