+++ src/routers/individualStatementRouter.ts
@@ -1,10 +1,6 @@
 import { NextFunction, Request, Response, Router } from "express";
-import { PrefixedUrls } from "../constants";
 import { handleExceptions } from "../utils/asyncHandler";
-import { selectLang } from "../utils/localise";
-import { addSearchParams } from "../utils/queryParams";
-import { getUrlWithTransactionIdAndSubmissionId } from "../utils/url";
-import { IndividualStatementHandler } from "./handlers/individual-statement/individualStatement";
+import { IndividualStatementHandler } from "./handlers/individual-statement/individualStatementHandler";
 
 const individualStatementRouter: Router = Router({ mergeParams: true });
 
@@ -16,11 +12,7 @@ individualStatementRouter.get("/", handleExceptions(async (req: Request, res: Re
 
 individualStatementRouter.post("/", handleExceptions(async (req: Request, res: Response, _next: NextFunction) => {
     const handler = new IndividualStatementHandler();
-    await handler.executePost(req, res);
-
-    const nextPageUrl = getUrlWithTransactionIdAndSubmissionId(PrefixedUrls.PSC_VERIFIED, req.params.transactionId, req.params.submissionId);
-    const lang = selectLang(req.query.lang);
-    res.redirect(addSearchParams(nextPageUrl, { lang }));
+    res.redirect(await handler.executePost(req, res));
 }));
 
 export default individualStatementRouter;
